♻️ Refactor: 리베이스 후 아바타 사이즈 상수 분리하여 호출방식으로 리팩토링 (#42)
