+++ src/components/avatar/AvatarSize.ts
@@ -1,0 +1,8 @@
+export const AVATAR_SIZE = {
+  xs: 'size-6',
+  sm: 'size-8',
+  md: 'size-10',
+  lg: 'size-12',
+  xl: 'size-16',
+} as const
+// 상수는 절대경로 사용이 권장되지 않기에 index.tsx 설정은 제외함.
