+++ src/components/index.tsx
@@ -12,12 +12,10 @@ import ToastContainer from '@/components/common/Toast/ToastContainer'
 import RootLayout from '@/components/layout/RootLayout'
 import MyPageLayout from '@/components/layout/MyPageLayout'
 import Chat from '@/components/chat/Chat'
-
 import ListItemSkeleton from '@/components/common/Skeleton/ListItemSkeleton'
 import ImageCardSkeleton from '@/components/common/Skeleton/ImageCardSkeleton'
 import AvatarSkeleton from '@/components/common/Skeleton/AvatarSkeleton'
-
-import Avatar from '@/components/common/Avatar'
+import Avatar from '@/components/avatar/Avatar'
 
 export {
   Badge,
